+++ Changelog.md
@@ -7,6 +7,7 @@
 ### Fixes
 * Fix crash while parsing `page_borders`
 * Fix parsing nil borders
+* Fix parsing document background
 
 ## 0.1.1 (2016-05-17)